Move container linux controller
