+++ api/pkg/controller/user-cluster-controller-manager/container-linux/README.md
@@ -1,0 +1,6 @@
+# `container-linux` contoller
+
+Responsibilites:
+
+* Label all container-linux-bases nodes with a `kubernetes.io/uses-container-linux: true` label
+* Deploy the [Container Linux Update Operator](https://github.com/coreos/container-linux-update-operator) operator and DaemonSet
